+++ CHANGELOG.md
@@ -16,7 +16,8 @@
 
 ### Fixes:
 
-* Fix 2fa route on webauthn page 
+* Fix relationship create and destroy with partial contact
+* Fix 2fa route on webauthn page
 * Fix tooltip on favorite icon
 * Fix icons disappeared on contact information
 * Fix CSV uploads with weird photo files